go to payment option should show the total amount on sale
Former-commit-id: 7f33e3e846e054569e31a5d58498cf83cea7dc19

+++ lib/domain/redux/app_actions/actions.dart
@@ -195,8 +195,8 @@ class CartQuantity {
 @immutable
 class SavePayment {
   final String note;
-  final int cashReceived;
-  final int customerChangeDue;
+  final double cashReceived;
+  final double customerChangeDue;
 
   SavePayment({this.customerChangeDue, this.note, this.cashReceived});
 }
